backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-api-python-client==2.144.0
apscheduler==3.10.4
python-multipart==0.0.12
pandas==2.2.2
numpy==1.26.4

